feat-mod: fund wallet: organized folder structure
