+++ server/interfaces/index.ts
@@ -1,0 +1,8 @@
+﻿import {TransactionStatus, WalletType, FundDataType, TransferDataType} from "./transactionsDataType";
+
+export {
+    TransactionStatus,
+    WalletType,
+    FundDataType,
+    TransferDataType
+}
