+++ server/interfaces/transactionsDataType.ts
