+++ server/interfaces/userDataType.ts
